backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.25.2
beautifulsoup4==4.12.3
openai==1.51.0
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.9.2
apscheduler==3.10.4
playwright==1.49.1
pdfplumber==0.11.0
supabase==2.10.0
websockets==14.1
python-multipart==0.0.9
